requirements.txt
argostranslate>=1.9.0
Pillow>=10.0.0
pywebview>=6.0
uiautomation>=2.0.29
